requirements.txt
attrs==23.2.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.2.0
cycler==0.12.1
Flask==3.0.2
flatbuffers==24.3.7
fonttools==4.50.0
humanfriendly==10.0
idna==3.6
imageio==2.34.0
itsdangerous==2.1.2
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lazy_loader==0.3
llvmlite==0.42.0
MarkupSafe==2.1.5
matplotlib==3.8.3
mpmath==1.3.0
networkx==3.2.1
numba==0.59.1
numpy==1.26.4
onnxruntime==1.17.1
opencv-python-headless==4.9.0.80
packaging==24.0
pillow==10.2.0
platformdirs==4.2.0
pooch==1.8.1
protobuf==5.26.0
PyMatting==1.1.12
pyparsing==3.1.2
pyreadline3==3.4.1
python-dateutil==2.9.0.post0
referencing==0.34.0
rembg==2.0.56
requests==2.31.0
rpds-py==0.18.0
scikit-image==0.22.0
scipy==1.12.0
six==1.16.0
sympy==1.12
tifffile==2024.2.12
tqdm==4.66.2
urllib3==2.2.1
Werkzeug==3.0.1
